+++ Frends.SFTP.DownloadFiles/Frends.SFTP.DownloadFiles.Tests/ConnectivityTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Linq;
 using System.Threading;
 using NUnit.Framework;
 using Frends.SFTP.DownloadFiles.Definitions;
@@ -47,7 +48,7 @@ namespace Frends.SFTP.DownloadFiles.Tests
         {
             var connection = Helpers.GetSftpConnection();
             connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyFile;
-            connection.PrivateKeyFilePassphrase = "passphrase";
+            connection.PrivateKeyPassphrase = "passphrase";
             connection.PrivateKeyFile = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key");
 
             var result = SFTP.DownloadFiles(_source, _destination, connection, _options, _info, new CancellationToken());
@@ -61,8 +62,9 @@ namespace Frends.SFTP.DownloadFiles.Tests
             var key = File.ReadAllText(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key"));
 
             var connection = Helpers.GetSftpConnection();
+            connection.HostKeyAlgorithm = HostKeyAlgorithms.RSA;
             connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyString;
-            connection.PrivateKeyFilePassphrase = "passphrase";
+            connection.PrivateKeyPassphrase = "passphrase";
             connection.PrivateKeyString = key;
 
             var result = SFTP.DownloadFiles(_source, _destination, connection, _options, _info, new CancellationToken());
@@ -79,6 +81,71 @@ namespace Frends.SFTP.DownloadFiles.Tests
             var result = SFTP.DownloadFiles(_source, _destination, connection, _options, _info, new CancellationToken());
             Assert.IsTrue(result.Success);
             Assert.AreEqual(1, result.SuccessfulTransferCount);
+        }
+
+        [Test]
+        public void DownloadFiles_TestWithInteractiveKeyboardAuthenticationAndPrivateKey()
+        {
+            var connection = Helpers.GetSftpConnection();
+            connection.Authentication = AuthenticationType.UsernamePrivateKeyFile;
+            connection.PrivateKeyFile = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key");
+            connection.Password = null;
+            connection.PrivateKeyPassphrase = "passphrase";
+            connection.UseKeyboardInteractiveAuthentication = true;
+            connection.PromptAndResponse = new PromptResponse[] { new PromptResponse { Prompt = "Password", Response = "pass" } };
+
+            var result = SFTP.DownloadFiles(_source, _destination, connection, _options, _info, new CancellationToken());
+            Assert.IsTrue(result.Success);
+            Assert.AreEqual(1, result.SuccessfulTransferCount);
+
+            connection.Authentication = AuthenticationType.UsernamePrivateKeyString;
+            connection.PrivateKeyFile = null;
+            connection.PrivateKeyString = File.ReadAllText(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key"));
+            connection.PrivateKeyPassphrase = "passphrase";
+
+            var destination = new Destination
+            {
+                Directory = Path.Combine(_workDir, "destination"),
+                Action = DestinationAction.Overwrite,
+                FileNameEncoding = FileEncoding.UTF8,
+                EnableBomForFileName = true
+            };
+
+            result = SFTP.DownloadFiles(_source, destination, connection, _options, _info, new CancellationToken());
+            Assert.IsTrue(result.Success);
+            Assert.AreEqual(1, result.SuccessfulTransferCount);
+        }
+
+        [Test]
+        public void DownloadFiles_TestShouldThrowWithoutPromptAndResponse()
+        {
+            var connection = Helpers.GetSftpConnection();
+            connection.Authentication = AuthenticationType.UsernamePrivateKeyFile;
+            connection.PrivateKeyFile = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key");
+            connection.Password = null;
+            connection.PrivateKeyPassphrase = "passphrase";
+            connection.UseKeyboardInteractiveAuthentication = true;
+            connection.HostKeyAlgorithm = HostKeyAlgorithms.RSA;
+            connection.ServerFingerPrint = "NUfXVu2omU2k3ELtmCzhkcERRLHAEbNakrpBgEXn8JM";
+
+            var ex = Assert.Throws<Exception>(() => SFTP.DownloadFiles(_source, _destination, connection, _options, _info, new CancellationToken()));
+            Assert.IsTrue(ex.Message.StartsWith("SFTP transfer failed: Failure in Keyboard-interactive authentication: No response given for server prompt request --> Password"));
+            
+            connection.Authentication = AuthenticationType.UsernamePrivateKeyString;
+            connection.PrivateKeyFile = null;
+            connection.PrivateKeyString = File.ReadAllText(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key"));
+            connection.PrivateKeyPassphrase = "passphrase";
+
+            var destination = new Destination
+            {
+                Directory = Path.Combine(_workDir, "destination"),
+                Action = DestinationAction.Overwrite,
+                FileNameEncoding = FileEncoding.UTF8,
+                EnableBomForFileName = true
+            };
+
+            ex = Assert.Throws<Exception>(() => SFTP.DownloadFiles(_source, destination, connection, _options, _info, new CancellationToken()));
+            Assert.IsTrue(ex.Message.StartsWith("SFTP transfer failed: Failure in Keyboard-interactive authentication: No response given for server prompt request --> Password"));
         }
 
         [Test]